+++ cmd/keytransparency-server/main.go
@@ -30,11 +30,11 @@ import (
 	"github.com/google/keytransparency/core/mutator/entry"
 	"github.com/google/keytransparency/impl/authentication"
 	"github.com/google/keytransparency/impl/authorization"
-	"github.com/google/keytransparency/impl/sql/directory"
-	"github.com/google/keytransparency/impl/sql/mutationstorage"
+	"github.com/google/keytransparency/impl/mysql/directory"
+	"github.com/google/keytransparency/impl/mysql/mutationstorage"
 
 	pb "github.com/google/keytransparency/core/api/v1/keytransparency_go_proto"
-	ktsql "github.com/google/keytransparency/impl/sql"
+	ktsql "github.com/google/keytransparency/impl/mysql"
 	grpc_middleware "github.com/grpc-ecosystem/go-grpc-middleware"
 	grpc_auth "github.com/grpc-ecosystem/go-grpc-middleware/auth"
 	grpc_prometheus "github.com/grpc-ecosystem/go-grpc-prometheus"
