+++ impl/mysql/directory/storage_test.go
@@ -22,7 +22,7 @@ import (
 	"github.com/golang/protobuf/proto"
 	"github.com/google/go-cmp/cmp"
 	"github.com/google/keytransparency/core/directory"
-	"github.com/google/keytransparency/impl/sql/testdb"
+	"github.com/google/keytransparency/impl/mysql/testdb"
 	"github.com/google/trillian/crypto/keyspb"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
